requirements.txt
altgraph>=0.17.3
certifi>=2022.12.7
charset-normalizer>=3.0.1
colorama>=0.4.6
dukpy>=0.3.0
et-xmlfile>=1.1.0
filelock>=3.9.0
future>=0.18.3
idna>=3.4
importlib-metadata>=6.0.0
jaro-winkler>=2.0.3
lief>=0.12.3
mysql-connector>=2.2.9
openpyxl>=3.1.0
packaging>=23.0
pefile>=2022.5.30
pyinstaller>=5.7.0
pyinstaller-hooks-contrib>=2022.15
pyodbc>=4.0.35
PyPAC>=0.16.1
pyparsing>=3.0.9
pywin32-ctypes>=0.2.0
PyYAML>=6.0
requests>=2.28.2
requests-file>=1.5.1
six>=1.16.0
tldextract>=3.4.0
urllib3>=1.26.14
zipp>=3.12.0

Django~=5.0
chardet~=5.2.0
dnspython~=2.4.2